Move `creatureUtils` from "reducers/utils" to "utils"
