+++ src/__tests__/utils-test.ts
@@ -2,6 +2,10 @@ import * as assert from 'assert';
 import {describe, it} from 'mocha';
 
 import {
+  createCreature,
+  createJob,
+} from '../test-utils';
+import {
   BattleFieldElement,
   BattleFieldMatrix,
   Card,
@@ -10,6 +14,7 @@ import {
   areGlobalPositionsEqual,
   choiceElementsAtRandom,
   createBattleFieldMatrix,
+  creatureUtils,
   findBattleFieldElementByCreatureId,
   findBattleFieldElementsByDistance,
   findCardUnderCursor,
@@ -469,6 +474,63 @@ describe('utils', function() {
         }),
         undefined,
       )
+    })
+  })
+
+  describe('creatureUtils', function() {
+    describe('getAttackPower', function() {
+      const jobs = [
+        {
+          ...createJob(),
+          attackPower: 2,
+        },
+      ]
+
+      it('_attackPowerForTest が存在しているときはその値を優先して返す', function() {
+        const creature = {
+          ...createCreature(),
+          _attackPowerForTest: 99,
+        }
+        assert.strictEqual(creatureUtils.getAttackPower(creature, jobs), 99)
+      })
+    })
+
+    describe('getMaxLifePoints', function() {
+      const jobs = [
+        {
+          ...createJob(),
+          maxLifePoints: 2,
+        },
+      ]
+
+      it('_maxLifePointsForTest が存在しているときはその値を優先して返す', function() {
+        const creature = {
+          ...createCreature(),
+          _maxLifePointsForTest: 99,
+        }
+        assert.strictEqual(creatureUtils.getMaxLifePoints(creature, jobs), 99)
+      })
+    })
+
+    describe('updateLifePoints', function() {
+      const jobs = [
+        {
+          ...createJob(),
+          maxLifePoints: 2,
+        },
+      ]
+      const creature = {
+        ...createCreature(),
+        lifePoints: 2,
+      }
+
+      it('lifePoints は 0 未満にならない', function() {
+        assert.strictEqual(creatureUtils.updateLifePoints(creature, jobs, -3).lifePoints, 0)
+      })
+
+      it('lifePoints は maxLifePoints を超えない', function() {
+        assert.strictEqual(creatureUtils.updateLifePoints(creature, jobs, 1).lifePoints, 2)
+      })
     })
   })
 })
